+++ src/main/java/de/sakurajin/sakuralib/compat/CompatLoader.java
@@ -1,5 +1,7 @@
 package de.sakurajin.sakuralib.compat;
 
+import net.fabricmc.loader.api.FabricLoader;
+
 /**
  * This class is used to load the compat modules.
  * It is called from the mod initializer and simply calls the init functions of the compat modules.
@@ -7,6 +9,8 @@ package de.sakurajin.sakuralib.compat;
  */
 public class CompatLoader {
     public static void init(){
-
+        if (FabricLoader.getInstance().isModLoaded("patchouli")){
+            PatchouliCompat.init();
+        }
     }
 }