+++ src/v1_20_1/java/de/sakurajin/sakuralib/version_stability/world/StructurePoolBasedGenerator.java
@@ -1,0 +1,14 @@
+package de.sakurajin.sakuralib.version_stability.world;
+
+import net.minecraft.registry.entry.RegistryEntry;
+import net.minecraft.structure.pool.StructurePool;
+import net.minecraft.util.Identifier;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.Heightmap;
+import net.minecraft.world.gen.structure.Structure;
+
+import java.util.Optional;
+
+public class StructurePoolBasedGenerator extends net.minecraft.structure.pool.StructurePoolBasedGenerator{
+
+}