+++ src/v1_20_3/java/de/sakurajin/sakuralib/mixin/minecraft/compat/StructurePoolBasedGeneratorMixin.java
@@ -1,0 +1,27 @@
+package de.sakurajin.sakuralib.mixin.minecraft.compat;
+
+import de.sakurajin.sakuralib.version_stability.blocks.BrushableBlock;
+import net.minecraft.block.AbstractBlock;
+import net.minecraft.block.Block;
+import net.minecraft.registry.entry.RegistryEntry;
+import net.minecraft.sound.SoundEvent;
+import net.minecraft.structure.pool.StructurePool;
+import net.minecraft.structure.pool.StructurePoolBasedGenerator;
+import net.minecraft.structure.pool.alias.StructurePoolAliasLookup;
+import net.minecraft.util.Identifier;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.Heightmap;
+import net.minecraft.world.gen.structure.Structure;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+import java.util.Optional;
+
+@Mixin(StructurePoolBasedGenerator.class)
+public interface StructurePoolBasedGeneratorMixin {
+
+}